hız desteği geldi

+++ ActionSim/src/actionsim/Point.java
@@ -2,10 +2,10 @@ package actionsim;
 
 public class Point {
 
-	public int x;
-	public int y;
+	public float x;
+	public float y;
 
-	public Point(int x, int y) {
+	public Point(float x, float y) {
 
 		this.x = x;
 		this.y = y;
